Individiual Command Cooldowns Update

+++ src/modules/commands/fun/addquote.ts
@@ -9,6 +9,10 @@ import knex from "../../../../db/knex";
 import { Collection, Message, MessageReaction } from "discord.js";
 
 import { createMetadata } from "../help/createMetadata";
+import {
+    Cooldown,
+    setCooldown
+} from "../../../common/parsing/middleware/comandCooldown";
 
 const checkBalance = async (id: string) => {
     const balance = (await knex("user").where({ userid: id }))[0].balance;
@@ -39,14 +43,13 @@ export default new Command()
                 "Sends a request for adding a new quote for level up messages. Costs 10k, if it gets rejected you get 5k back, and if it doesn't get any response in 24h then you get your 10k back"
         })
     )
+    .use<Cooldown>((context, next) => {
+        setCooldown(context, next, 10000);
+    })
     .use<ParseArgumentsState>(async context => {
         const { message } = context;
         const { args } = context.state;
         try {
-            if (message.channel.type === "dm")
-                return message.channel.send(
-                    `Im sorry but you can't add a quote from dms anymore.`
-                );
             if (!args.length) {
                 return message.channel.send(
                     `**Error**: You cannot put an empty quote.`
@@ -152,7 +155,7 @@ export default new Command()
                         }
                     }
                 );
-                message.delete();
+                if (message.channel.type !== "dm") message.delete();
             }, 1000);
         } catch (e) {
             return message.channel.send("**Error**: Something went wrong.");
